requirements.txt
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
scikit-learn>=1.3
scipy>=1.10
torch>=2.2
torchvision>=0.17
tensorboard>=2.14
filelock>=3.12
wonderwords>=2.2
weightwatcher>=0.7

# Optional extras for non-default workflows:
# transformers>=4.40
# datasets>=2.18
# flwr-datasets>=0.5
# flamby>=0.1.1
